Everything works, just need the listing of houses to be displayed in the listing page and in the search for houses page.

+++ RealEstateTermProject/SoapUserFunc.asmx.cs
@@ -125,7 +125,7 @@ namespace RealEstateTermProject
         }
 
         //adding to requested seller table from show real estate comapny
-        public void AddToRequestedSeller(string username, int userID, string address, string amentities, string state, int price, string city, string garage,
+        public void AddToRequestedSeller(string username, int userID, string address, string amentities, string state, decimal price, string city, string garage,
            string description, int homesize, string houseImages, int year, int bathroom, int bedroom, string proptype, string utils, int AgentId)
         {
             SqlCommand objCommand = new SqlCommand();
